client: make prom client optional
