+++ packages/client/package.json
@@ -86,7 +86,6 @@
     "kzg-wasm": "^0.4.0",
     "level": "^8.0.0",
     "memory-level": "^1.0.0",
-    "prom-client": "^15.1.0",
     "verkle-cryptography-wasm": "^0.4.2",
     "winston": "^3.3.3",
     "winston-daily-rotate-file": "^4.5.5",
@@ -106,6 +105,9 @@
     "testdouble": "^3.16.6",
     "testdouble-timers": "^0.1.1",
     "ws": "^8.14.2"
+  },
+  "optionalDependencies": {
+    "prom-client": "^15.1.0"
   },
   "engines": {
     "node": ">=18"
