+++ packages/client/src/config.ts
@@ -9,7 +9,8 @@ import { Event, EventBus } from './types.js'
 import { isBrowser, short } from './util/index.js'
 
 import type { Logger } from './logging.js'
-import type { EventBusType, MultiaddrLike, PrometheusMetrics } from './types.js'
+import type { EventBusType, MultiaddrLike } from './types.js'
+import type { PrometheusMetrics } from './util/metrics.js'
 import type { BlockHeader } from '@ethereumjs/block'
 import type { VM, VMProfilerOpts } from '@ethereumjs/vm'
 import type { Multiaddr } from '@multiformats/multiaddr'
